float --> double, moved vectors and matrices into their own namespace, extended vector functions, added unit test for Vector3D
git-svn-id: https://aquitaine.dnsalias.net:8443/svn/mathlib2/trunk@5 35ab40ff-a1d5-ea4a-9275-db5122b6521a

+++ Library/BasicMatrix.cs
@@ -1,8 +1,14 @@
 namespace Library
 {
+	/// <summary>
+	/// A matrix base class
+	/// </summary>
 	public abstract class BasicMatrix
 	{
-		internal float[,] Cell;
+		/// <summary>
+		/// The cell array
+		/// </summary>
+		internal double[,] Cell;
 
 		/// <summary>
 		/// Sets the matrix to an identity matrix
@@ -13,6 +19,6 @@ namespace Library
 		/// Assigns matrix values
 		/// </summary>
 		/// <param name="value">Value to fill in</param>
-		public abstract void Fill(float value);
+		public abstract void Fill(double value);
 	}
 }
