+++ Library/Interpolation.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Library.Vector;
 
 namespace Library
 {
@@ -17,9 +18,9 @@ namespace Library
 		/// <param name="p3">p3</param>
 		/// <param name="t">positional weighting factor</param>
 		/// <returns>Interpolated vector</returns>
-		public static Vector3D CatmullRom(Vector3D p0, Vector3D p1, Vector3D p2, Vector3D p3, float t)
+		public static Vector3D CatmullRom(Vector3D p0, Vector3D p1, Vector3D p2, Vector3D p3, double t)
 		{
-			float t2 = t * t, t3 = t2 * t;
+			double t2 = t * t, t3 = t2 * t;
 			Vector3D ret = 0.5f * ((2f * p1) +
 							(-p0 + p2) * t +
 							(2f * p0 - 5f * p1 + 4f * p2 - p3) * t2 +
@@ -36,10 +37,10 @@ namespace Library
 		/// <param name="p3">p3</param>
 		/// <param name="t">positional weighting factor</param>
 		/// <returns>Interpolated vector</returns>
-		public static Vector3D CubicBezier(Vector3D p0, Vector3D p1, Vector3D p2, Vector3D p3, float t)
+		public static Vector3D CubicBezier(Vector3D p0, Vector3D p1, Vector3D p2, Vector3D p3, double t)
 		{
-			float t2 = t * t, t3 = t2 * t;
-			float b = 1f - t, b2 = b * b, b3 = b2 * b;
+			double t2 = t * t, t3 = t2 * t;
+			double b = 1f - t, b2 = b * b, b3 = b2 * b;
 			return (p0 * b3 + 3f * p1 * b2 * t + 3f * p2 * b * t2 + p3 * t3);
         }
 
@@ -56,9 +57,9 @@ namespace Library
         /// <param name="p3">p3</param>
         /// <param name="t">positional weighting factor</param>
         /// <returns>Interpolated vector</returns>
-        public static Vector2D CatmullRom(Vector2D p0, Vector2D p1, Vector2D p2, Vector2D p3, float t)
+        public static Vector2D CatmullRom(Vector2D p0, Vector2D p1, Vector2D p2, Vector2D p3, double t)
         {
-            float t2 = t * t, t3 = t2 * t;
+            double t2 = t * t, t3 = t2 * t;
             Vector2D ret = 0.5f * ((2f * p1) +
                             (-p0 + p2) * t +
                             (2f * p0 - 5f * p1 + 4f * p2 - p3) * t2 +
@@ -75,10 +76,10 @@ namespace Library
         /// <param name="p3">p3</param>
         /// <param name="t">positional weighting factor</param>
         /// <returns>Interpolated vector</returns>
-        public static Vector2D CubicBezier(Vector2D p0, Vector2D p1, Vector2D p2, Vector2D p3, float t)
+        public static Vector2D CubicBezier(Vector2D p0, Vector2D p1, Vector2D p2, Vector2D p3, double t)
         {
-            float t2 = t * t, t3 = t2 * t;
-            float b = 1f - t, b2 = b * b, b3 = b2 * b;
+            double t2 = t * t, t3 = t2 * t;
+            double b = 1f - t, b2 = b * b, b3 = b2 * b;
             return (p0 * b3 + 3f * p1 * b2 * t + 3f * p2 * b * t2 + p3 * t3);
         }
