+++ Library/Sphere3D.cs
@@ -1,20 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Library.Vector;
 
 namespace Library
 {
 	public class Sphere3D
 	{
 		public Vector3D Center;
-		public float Radius;
+		public double Radius;
 
 		/// <summary>
 		/// Creates a sphere based on a point and a radius
 		/// </summary>
 		/// <param name="center">Center point</param>
 		/// <param name="radius">Radius</param>
-		public Sphere3D(Vector3D center, float radius)
+		public Sphere3D(Vector3D center, double radius)
 		{
 			Center = center;
 			Radius = radius;
@@ -28,7 +29,7 @@ namespace Library
 		public bool Intersects(Line3D line)
 		{
 			// Get distance to center
-			float distance = line.GetDistance(Center);
+			double distance = line.GetDistance(Center);
 			// If the distance is smaller than the radius, we are cutting the edge
 			if (distance <= Radius) return true;
 			return false;
@@ -42,7 +43,7 @@ namespace Library
 		public bool Intersects(LineSegment3D line)
 		{
 			// Get distance to center
-			float distance = line.GetDistance(Center);
+			double distance = line.GetDistance(Center);
 			// If the distance is smaller than the radius, we are cutting the edge
 			if (distance <= Radius) return true;
 			return false;
@@ -52,13 +53,13 @@ namespace Library
 		/// Returns the distance to the given sphere
 		/// </summary>
 		/// <param name="sphere">The sphere to test</param>
-		/// <returns>Float</returns>
-		public float GetDistance(Sphere3D sphere)
+		/// <returns>double</returns>
+		public double GetDistance(Sphere3D sphere)
 		{
 			// Get difference vector
 			Vector3D difference = sphere.Center - this.Center;
 			// Get value
-			float distance = difference.Magnitude() - sphere.Radius - this.Radius;
+			double distance = difference.Magnitude() - sphere.Radius - this.Radius;
 			// Return distance
 			if (distance < 0f) return 0f;
 			return distance;
